+++ Cargo.toml
@@ -1,6 +1,11 @@
 [workspace]
 resolver = "2"
-members = ["nym-vpn-lib", "nym-vpn-cli", "nym-vpnd"]
+members = [
+    "crates/nym-gateway-directory",
+    "nym-vpn-cli",
+    "nym-vpn-lib",
+    "nym-vpnd",
+]
 exclude = ["nym-vpn-desktop/src-tauri"]
 
 # For local development
@@ -18,6 +23,49 @@ exclude = ["nym-vpn-desktop/src-tauri"]
 # nym-validator-client = { path = "../nym/common/client-libs/validator-client" }
 # nym-wireguard-types = { path = "../nym/common/wireguard-types" }
 
+[workspace.package]
+authors = ["Nym Technologies SA"]
+repository = "https://github.com/nymtech/nym-vpn-client"
+homepage = "https://nymtech.net"
+documentation = "https://nymtech.net"
+edition = "2021"
+
 [workspace.dependencies]
 tokio = { version = "1.8" }
+rand = "0.7.3"
 
+anyhow = "1.0.71"
+bincode = "1.3.3"
+bytes = "1.0"
+futures = "0.3.15"
+ipnetwork = "0.16"
+pnet_packet = "0.34.0"
+itertools = "0.12.1"
+lazy_static = "1.4.0"
+log = "0.4.20"
+netdev = "0.24.0"
+serde = "1.0.192"
+serde_json = "1.0.91"
+signature = "1"
+tap = "1.0.1"
+thiserror = "1.0.38"
+tokio-util = { version = "0.7.10", features = ["codec"] }
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tun2 = { version = "1.2.3", features = ["async"] }
+uniffi = { version = "0.26.1", features = ["cli"] }
+url = "2.4"
+hickory-resolver = "0.24.0"
+
+nym-bin-common = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-client-core = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-config = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-crypto = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-explorer-client = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-ip-packet-requests = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-node-requests = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-sdk = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-task = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-topology = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-validator-client = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
+nym-wireguard-types = { git = "https://github.com/nymtech/nym", rev = "f5378e8a868d9f99adf61b0aa018593e32bc03de" }
