+++ nym-vpn-lib/src/lib.rs
@@ -5,13 +5,13 @@ uniffi::setup_scaffolding!();
 
 use crate::config::WireguardConfig;
 use crate::error::{Error, Result};
-use crate::gateway_client::{Config, GatewayClient, IpPacketRouterAddress};
+use nym_gateway_directory::{Config, GatewayClient, IpPacketRouterAddress};
 use crate::mixnet_connect::setup_mixnet_client;
 use crate::tunnel::{setup_route_manager, start_tunnel, Tunnel};
 use crate::util::{handle_interrupt, wait_for_interrupt};
 use crate::wg_gateway_client::{WgConfig, WgGatewayClient};
 use futures::channel::{mpsc, oneshot};
-use gateway_client::{EntryPoint, ExitPoint};
+use nym_gateway_directory::{EntryPoint, ExitPoint};
 use log::{debug, error, info};
 use mixnet_connect::SharedMixnetClient;
 use nym_task::TaskManager;
@@ -43,7 +43,6 @@ use tun2::AsyncDevice;
 pub mod config;
 mod connection_monitor;
 pub mod error;
-pub mod gateway_client;
 mod icmp_connection_beacon;
 pub mod mixnet_connect;
 pub mod mixnet_connection_beacon;
